four phases of Employee of the Planet

+++ src/main/java/satisfied/ItemFlow.java
@@ -2,10 +2,10 @@ package satisfied;
 
 public class ItemFlow {
   public final Item item;
-  public final int count;
+  public final double count;
   public final double rate;
 
-  public ItemFlow(Item item, int count, double rate) {
+  public ItemFlow(Item item, double count, double rate) {
     this.item = item;
     this.count = count;
     this.rate = rate;
